+++ packages/wallet-sdk/src/mocks/relay.ts
@@ -1,9 +1,9 @@
 import { AddressString, HexString, ProviderType } from '../core/type';
-import { Web3Method as SupportedWeb3Method, Web3Method } from '../core/type/Web3Method';
-import { Web3Response } from '../core/type/Web3Response';
 import { ScopedLocalStorage } from '../lib/ScopedLocalStorage';
 import { RelayAbstract } from '../relay/RelayAbstract';
 import { Session } from '../relay/Session';
+import { Web3Method as SupportedWeb3Method, Web3Method } from '../relay/walletlink/type/Web3Method';
+import { Web3Response } from '../relay/walletlink/type/Web3Response';
 import { MOCK_ADDERESS, MOCK_TX } from './fixtures';
 
 function makeMockReturn<T extends SupportedWeb3Method>(response: Web3Response<T>) {
