+++ packages/wallet-sdk/src/provider/Web3Provider.ts
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, version 2.0
 
 import { Callback } from '../core/type';
-import { JSONRPCRequest, JSONRPCResponse } from '../core/type/JSONRPC';
+import { JSONRPCRequest, JSONRPCResponse } from '../provider/JSONRPC';
 
 export interface Web3Provider {
   send(request: JSONRPCRequest): JSONRPCResponse;
