+++ packages/wallet-sdk/src/relay/RelayAbstract.ts
@@ -1,10 +1,10 @@
 import { ErrorHandler, serializeError, standardErrors } from '../core/error';
 import { AddressString, IntNumber, ProviderType, RegExpString } from '../core/type';
-import { EthereumTransactionParams } from '../core/type/EthereumTransactionParams';
-import { JSONRPCRequest, JSONRPCResponse } from '../core/type/JSONRPC';
-import { SupportedWeb3Method, Web3Request } from '../core/type/Web3Request';
-import { Web3Response } from '../core/type/Web3Response';
+import { JSONRPCRequest, JSONRPCResponse } from '../provider/JSONRPC';
 import { Session } from './Session';
+import { EthereumTransactionParams } from './walletlink/type/EthereumTransactionParams';
+import { SupportedWeb3Method, Web3Request } from './walletlink/type/Web3Request';
+import { Web3Response } from './walletlink/type/Web3Response';
 
 export const WALLET_USER_NAME_KEY = 'walletUsername';
 export const LOCAL_STORAGE_ADDRESSES_KEY = 'Addresses';
