+++ packages/wallet-sdk/src/relay/mobile/MobileRelay.ts
@@ -1,8 +1,8 @@
-import { Web3Request } from '../../core/type/Web3Request';
-import { Web3Response } from '../../core/type/Web3Response';
 import { getLocation } from '../../core/util';
 import { CancelablePromise } from '../RelayAbstract';
 import { WalletLinkResponseEventData } from '../walletlink/type/WalletLinkEventData';
+import { Web3Request } from '../walletlink/type/Web3Request';
+import { Web3Response } from '../walletlink/type/Web3Response';
 import { WalletLinkRelay, WalletLinkRelayOptions } from '../walletlink/WalletLinkRelay';
 import { MobileRelayUI } from './MobileRelayUI';
