+++ packages/wallet-sdk/src/relay/walletlink/type/WalletLinkEventData.ts
@@ -1,8 +1,8 @@
 // Copyright (c) 2018-2023 Coinbase, Inc. <https://www.coinbase.com/>
 // Licensed under the Apache License, version 2.0
 
-import { Web3Request } from '../../../core/type/Web3Request';
-import { Web3Response } from '../../../core/type/Web3Response';
+import { Web3Request } from './Web3Request';
+import { Web3Response } from './Web3Response';
 
 type Type =
   | 'SESSION_ID_REQUEST'
