+++ packages/wallet-sdk/src/relay/walletlink/ui/WalletLinkRelayUI.ts
@@ -1,8 +1,8 @@
 import { ErrorHandler } from '../../../core/error';
-import {} from '../../../core/type/Web3Request';
-import {} from '../../../core/type/Web3Response';
 import { injectCssReset } from '../../../lib/cssReset';
 import { RelayUI, RelayUIOptions } from '../../RelayUI';
+import {} from '../type/Web3Request';
+import {} from '../type/Web3Response';
 import { LinkFlow } from './components/LinkFlow/LinkFlow';
 import { Snackbar, SnackbarInstanceProps } from './components/Snackbar/Snackbar';
